refactor: deduplicate TapToDismiss into one shared component
