+++ src/screens/change-wallet/ChangeWalletSheet.tsx
@@ -12,7 +12,8 @@ import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { type MenuConfig, type MenuItem } from '@/components/DropdownMenu';
 import { EasingGradient } from '@/components/easing-gradient/EasingGradient';
 import { SheetHandleFixedToTop } from '@/components/sheet';
-import { Panel, TapToDismiss } from '@/components/SmoothPager/ListPanel';
+import { Panel } from '@/components/SmoothPager/ListPanel';
+import { TapToDismiss } from '@/components/TapToDismiss';
 import { FeatureHintTooltip, type TooltipRef } from '@/components/tooltips/FeatureHintTooltip';
 import { Box, globalColors, HitSlop, Inline, Text } from '@/design-system';
 import type { EthereumAddress } from '@/entities/wallet';
